+++ code/PaperTrader/core/models/order.h
@@ -6,11 +6,11 @@ struct Order {
     int id = 0;
     QString symbol;
     double price = 0.0;
-    int quantity = 0;
+    double quantity = 0.0;
     QString side;     // "BUY" / "SELL"
     QString type;     // "Market", "Limit", etc.
     QString status;   // "Open", "Filled", "Cancelled"
     QDateTime timestamp;
     double filledPrice = 0.0;
-    int filledQuantity = 0;
+    double filledQuantity = 0.0;
 };
